requirements-remote.txt
# Primary
Django==2.2.4
django-admin-env-notice==0.2
Pillow==6.1.0
psycopg2==2.7
celery==4.3.0
django-redis==4.10.0
redis==3.3.5

# Production
gunicorn==19.9
raven==6.10.0

# Tests / linters
pycodestyle==2.5.0
coverage==4.5.3
coverage-badge==1.0.1
pylama==7.7.1
factory-boy==2.12.0

# Performance
django-debug-toolbar==2.0
